feat: integrate the workflow visualizer in the show page
